+++ test/special-browsers.test.ts
@@ -18,13 +18,57 @@ function makeApp (middleware: ReturnType<typeof csp>) {
 }
 
 describe('special browsers', () => {
+  describe('Firefox 4', () => {
+    const userAgentString = AGENTS['Firefox 4'].string;
+
+    it('sets the header properly', async () => {
+      const app = makeApp(csp({
+        directives: {
+          defaultSrc: ["'self'"],
+          connectSrc: ['connect.com'],
+        },
+      }));
+
+      const res = await request(app)
+        .get('/')
+        .set('User-Agent', userAgentString);
+
+      expect(parseCsp(res.header['x-content-security-policy'])).toStrictEqual({
+        allow: ["'self'"],
+        'xhr-src': ['connect.com'],
+      });
+    });
+
+    it("doesn't mutate directives for future requests (tests an old bug)", async () => {
+      const app = makeApp(csp({
+        directives: {
+          defaultSrc: ["'self'"],
+          connectSrc: ['connect.com'],
+        },
+      }));
+
+      await request(app)
+        .get('/')
+        .set('User-Agent', userAgentString);
+
+      const res = await request(app)
+        .get('/')
+        .set('User-Agent', AGENTS['Firefox 53'].string);
+
+      expect(parseCsp(res.header['content-security-policy'])).toStrictEqual({
+        'default-src': ["'self'"],
+        'connect-src': ['connect.com'],
+      });
+    });
+  });
+
   [
     'Firefox 22',
     'Firefox OS 1.4',
     'Firefox for Android 24',
   ].forEach((browser) => {
     describe(browser, () => {
-      it('sets the header properly', (done) => {
+      it('sets the header properly', async () => {
         const app = makeApp(csp({
           directives: {
             defaultSrc: ["'self'"],
@@ -32,20 +76,14 @@ describe('special browsers', () => {
           },
         }));
 
-        request(app).get('/').set('User-Agent', AGENTS[browser as keyof typeof AGENTS].string)
-          .end((err, res) => {
-            if (err) {
-              done(err);
-              return;
-            }
+        const res = await request(app)
+          .get('/')
+          .set('User-Agent', AGENTS[browser as keyof typeof AGENTS].string);
 
-            expect(parseCsp(res.header['x-content-security-policy'])).toStrictEqual({
-              'default-src': ["'self'"],
-              'xhr-src': ['connect.com'],
-            });
-
-            done();
-          });
+        expect(parseCsp(res.header['x-content-security-policy'])).toStrictEqual({
+          'default-src': ["'self'"],
+          'xhr-src': ['connect.com'],
+        });
       });
     });
   });