+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+### Fixed
+- Fixed a bug where a request from Firefox 4 could delete `default-src` from future responses
+
 ## 2.9.1 - 2019-09-04
 ### Changed
 - Updated `bowser` subdependency from 2.5.3 to 2.5.4. See [#88](https://github.com/helmetjs/csp/pull/88)